Removed classes from main code and made their own files
also added some bug handling in wave_resistance.elevation_terms in handing small values of sech(KH)
